requirements.txt
defusedxml==0.4.1
Django==1.9.9
django-allauth==0.27.0
django-environ==0.4.0
django-rest-auth==0.8.1
djangorestframework==3.4.6
oauthlib==1.1.2
psycopg2==2.6.2
python3-openid==3.0.10
requests==2.11.1
requests-oauthlib==0.6.2
six==1.10.0
defusedxml==0.4.1
Django==1.9.9
django-allauth==0.27.0
django-autoslug==1.9.3
django-countries==4.0
django-environ==0.4.0
django-model-utils==2.4
django-rest-auth==0.8.1
djangorestframework==3.4.6
oauthlib==1.1.2
psycopg2==2.6.2
python3-openid==3.0.10
requests==2.11.1
requests-oauthlib==0.6.2
six==1.10.0
defusedxml==0.4.1
Django==1.9.9
django-allauth==0.27.0
django-autoslug==1.9.3
django-countries==4.0
django-environ==0.4.0
django-model-utils==2.4
django-rest-auth==0.8.1
djangorestframework==3.4.6
oauthlib==1.1.2
Pillow==3.3.1
psycopg2==2.6.2
python3-openid==3.0.10
requests==2.11.1
requests-oauthlib==0.6.2
six==1.10.0
